Fixed problem where the system would look for a cutting tool by ID rather than by tool number when the operator changed the value in a drilling cycle object.
Changed option so that an operator now may change the cutting tool number to zero (or negative)
for a drilling cycle object so as to disassociate it from any cutting tools.

Added check when changing a cutting tool number in a drilling cycle object so that, if a positive
number is entered, the cutting tool is found.  If not, an error is given and the change
is aborted.