Changed iso_codes.py to define methods rather than values.  This allows other classes to overload those methods with their own definitions.  I did this so that we could have spaces in the GCode generated for EMC2-based machines.
Changed emc2.py to overload methods in both the iso_codes.py and the iso.py files so that the resultant GCode is much more voerbose and, hopefully, more readable.

Changed the hm50.py and gantry_router.py class definitions to be based on the emc2 class rather than going directly to the iso and iso_codes classes.  This allows the emc2 GCode overloads to be verbose (because EMC2 can handle it) and the hm50 and gantry_router definitions to overload ONLY those methods that are specific to their machine.

Changed the iso_with_spaces.py routine so that it overloads the iso_codes.py methods rather than the iso.py methods.