Added an extra layer to the AppendTextToProgram() methods in some of the maching operations so that I could generate a pocket operation on the fly without having to add either it or its graphics to the main model.
I think I have most of the Inlay operation together now.  All I need to do now is be home early enough to be able to run the mill with the generated GCode to see if it actually works.  It now adds a pocket operation to cleanout the majority of material from inside the female half and then a chamfering bit to form the edges of both the female and male parts.