This is a little closer to getting the ramp entry move working but it's still not quite there.  Where a sketch contains a child object that is neither a line nor an arc, the code uses the BRepMesh::Mesh() routine to convert the element into small lines that don't devaite from the original curve by more than a 'tolerance' amount.  Unfortunately, this is working where we want to go all the way from the beginning to the end of this edge but not for when we only want to go part way along this edge.  When we get half way down the 'step-down' depth, we want to double back on ourselves so that we end up at our starting point at the step-down depth.  If the last edge to traverse before this 'half way down' point is along an edge that is neither a line nor an arc, we get the problem shown up.
I can see a routine to 'trim' an edge using a start and end 'U'
value but it returns a Handle_Adaptor3d_HCurve object and
I can't figure out what to do with that object so that I can
use the BRepMesh::Mesh() routine on it.

I know this is a long comment but I'm hoping someone will
take pity on me and tell me the answer.